Build: Win32 убран, общие настройки вынесены в Common.props
Пути к SFML, список библиотек и стандарт языка заданы один раз.
Убраны дубли NOMINMAX и предупреждение о size_t.

+++ XYZEngine/GameWorld.cpp
@@ -108,7 +108,7 @@ namespace XYZEngine
 	}
 	void GameWorld::Clear()
 	{
-		for (int i = gameObjects.size() - 1; i >= 0; i--)
+		for (int i = static_cast<int>(gameObjects.size()) - 1; i >= 0; i--)
 		{
 			if (gameObjects[i] == nullptr)
 			{
